refactor: move secondary icon button
