+++ packages/components/buttons/secondary-icon-button/src/secondary-icon-button.visualroute.js
@@ -4,7 +4,7 @@ import {
   InformationIcon,
   SecondaryIconButton,
 } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 export const routePath = '/secondary-icon-button';
