+++ packages/components/buttons/secondary-icon-button/src/secondary-icon-button.visualspec.js
